+++ src/Logfella/Adapters/HostBuilderExtensions.cs
@@ -1,3 +1,4 @@
+using Microsoft.Extensions.DependencyInjection.Extensions;
 using Microsoft.Extensions.Hosting;
 
 namespace Logfella.Adapters
@@ -5,6 +6,10 @@ namespace Logfella.Adapters
     public static class HostBuilderExtensions
     {
         public static IHostBuilder UseLogfella(this IHostBuilder builder) =>
-            builder.ConfigureLogging(b => b.UseLogfella());
+            builder
+                .ConfigureServices(
+                    (_, svc) => svc.TryAddTransient(
+                        _ => Log.LogWriter))
+                .ConfigureLogging(b => b.UseLogfella());
     }
 }