+++ src/Logfella/AspNetCore/ApplicationBuilderExtensions.cs
@@ -1,4 +1,6 @@
+using System;
 using Microsoft.AspNetCore.Builder;
+using Microsoft.AspNetCore.Http;
 
 namespace Logfella.AspNetCore
 {
@@ -6,10 +8,12 @@ namespace Logfella.AspNetCore
     {
         public static IApplicationBuilder UseRequestLogging(
             this IApplicationBuilder builder,
-            bool isEnabled = true)
-        {
+            bool isEnabled = true) =>
             builder.UseMiddleware<RequestLoggingMiddleware>(isEnabled);
-            return builder;
-        }
+
+        public static IApplicationBuilder UsePerRequestLogWriter(
+            this IApplicationBuilder builder,
+            Func<HttpContext, ILogWriter> logWriterFactory) =>
+            builder.UseMiddleware<PerRequestLogWriterMiddleware>(logWriterFactory);
     }
 }